- move dictdiffer out of lib, and then hide lib (gets used by venv apparently)
