+++ .gitignore
@@ -1,3 +1,5 @@
+lib/*
+
 # Created by https://www.gitignore.io/api/macos,python
 # Edit at https://www.gitignore.io/?templates=macos,python
